Microservices/CIConnector/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
PyYAML==6.0.2
ruamel.yaml==0.18.6
